requirements.txt
python-docx==0.8.11
requests>=2.32.3,<3
lxml>=5,<6
pdfplumber>=0.11.0,<0.12
pytesseract>=0.3.10,<0.4
pdf2image>=1.17.0,<1.18
Pillow>=10.0.0,<11
